requirements.txt
boto3==1.34.66
langchain==0.1.20
langchain-openai==0.1.5
python-dotenv==1.0.1
pydantic==2.6.4
fastapi==0.110.0
uvicorn==0.29.0
openai==1.14.1
numpy==1.26.4
tiktoken==0.9.0
click==8.1.8
tenacity==8.5.0
PyPDF2==3.0.1
python-multipart==0.0.9
python-docx==1.0.1
email-validator==2.1.1
supertokens-python==0.18.0